requirements.txt
appdirs==1.4.4
asn1crypto==1.5.1
async-generator==1.10
attrs==21.2.0
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binwalk==2.3.3
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
chepy==4.0.0
cheroot==8.6.0
CherryPy==18.6.1
cli-helpers==2.2.1
click==8.1.2
colorama==0.4.4
colorlog==6.6.0
configobj==5.0.6
cprint==1.2.2
crccheck==1.1
crit==0.0.1
crypto==1.4.1
cryptography==37.0.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
delegator.py==0.1.1
distro-info==1.1
dnspython==2.2.1
docstring-parser==0.14.1
emoji==1.7.0
enum34==1.1.10
evdev==1.6.0
exrex==0.10.5
fire==0.4.0
Flask==2.1.1
fonttools==4.32.0
frida==15.1.17
frida-tools==10.6.1
fs==2.4.12
gcloud==0.17.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
gyp==0.1
h11==0.13.0
hexdump==3.3
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
ipaddress==1.0.23
itsdangerous==2.1.2
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
javalang==0.13.0
Jinja2==3.1.1
Js2Py==0.71
jsonpickle==2.2.0
jws==0.1.3
kiwisolver==1.3.2
lazy-import==0.2.2
litecli==1.8.0
lxml==4.8.0
lz4==4.0.0+dfsg
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.5.1
more-itertools==8.12.0
mpmath==0.0.0
mysql-connector-python==8.0.30
mysql-connector=2.2.9
Naked==0.1.31
numpy==1.21.5
oauth2client==3.0.0
objection==1.11.0
olefile==0.46
outcome==1.2.0
packaging==21.3
Padding==0.5
paramiko==2.10.4
pbkdf2==1.3
pexpect==4.8.0
phpserialize==1.3
Pillow==9.0.1
pluginbase==1.0.1
portend==3.1.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycipher==0.5.2
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.4.3
pycryptodome-test-vectors==1.0.8
pycryptodomex==3.14.1
pycurl==7.44.1
pydash==5.1.0
pyelftools==0.28
Pygments==2.11.2
PyGObject==3.42.0
pyjsparser==2.7.1
PyJWT==1.7.1
PyNaCl==1.5.0
pynput==1.7.6
PyOpenGL==3.1.5
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
Pyrebase==3.0.27
PySocks==1.7.1
python-apt==2.3.0+b1
python-dateutil==2.8.1
python-editor==1.0.4
python-jwt==2.0.1
python-magic==0.4.25
python-xlib==0.31
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qark==4.0.0
reflutter==0.6.6
regex==2022.6.2
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
scipy==1.7.3
scp==0.14.4
selenium==4.3.0
semver==2.13.0
shellescape==3.8.1
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
SQLAlchemy==1.4.34
sqlparse==0.4.2
stegcracker==2.1.0
Sublist3r==1.0
sympy==1.9
tabulate==0.8.9
tempora==5.0.1
termcolor==1.1.0
tld==0.11.11
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ufoLib2==0.13.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.9
wapiti3==3.0.4
wcwidth==0.2.5
web.py==0.62
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.1
wsproto==1.1.0
yaswfp==0.9.3
zc.lockfile==2.0
zipp==3.8.0
